requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
scapy>=2.5.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fpdf2>=2.7.9
python-magic-bin>=0.4.14
chardet>=5.2.0
colorama>=0.4.6
matplotlib>=3.8.0
